NP-9074 Refactor resources actions and reducer to RTK
