+++ src/redux/reducers/rootReducer.ts
@@ -1,6 +1,6 @@
 import { Notification } from '../../types/notification.types';
 import { User } from '../../types/user.types';
-import { ResourceType } from '../actions/resourcesActions';
+import { ResourceType } from '../resourcesSlice';
 
 export interface RootStore {
   notification: Notification | null;
