+++ src/utils/hooks/useFetchResource.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useFetch } from './useFetch';
 import { RootStore } from '../../redux/reducers/rootReducer';
-import { ResourceType, setResource } from '../../redux/actions/resourcesActions';
+import { ResourceType, setResource } from '../../redux/resourcesSlice';
 
 // This hook is used to fetch all top-level institutions and put them in Redux, to avoid fetching same data many times
 export const useFetchResource = <T extends ResourceType>(
